Convert the *.sql files into Doctrines DBAL schema representation
1) The sql dumps of the following extensions were added:
* typo3/sysext/core
* typo3/sysext/extbase
* typo3/sysext/extensionmanager
* typo3/sysext/felogin
* typo3/sysext/filemetadata
* typo3/sysext/frontend
* typo3/sysext/impexp
* typo3/sysext/indexed_search
* typo3/sysext/indexed_search_mysql
* typo3/sysext/linkvalidator
* typo3/sysext/openid
* typo3/sysext/rsauth
* typo3/sysext/rtehtmlarea
* typo3/sysext/scheduler
* typo3/sysext/sys_action
* typo3/sysext/sys_note
* typo3/sysext/version
* typo3/sysext/workspaces

Migrated ext_tables_static+adt.sql of the extension manager to
DefaultData.php.

2) Adjust the core
The method \TYPO3\CMS\Install\Controller\Action\Step\DatabaseData::importDatabaseData()
was modified to be able for installing the database with the
Schema.php files instead of the *.sql dumps.

To get the expected schema it calls
\TYPO3\CMS\Install\Service\SqlExpectedSchemaServicegetTablesDefinitionAsDoctrineSchemaObjects()
which require the Schema.php files of  extension who provide those files.
To be available via the extension configuration, the file Schema.php was added to
\TYPO3\CMS\Core\Compatibility\LoadedExtensionArrayElement::$extensionFilesToCheckFor.

SQL dumps of some extensions just extend a table of another extension like
ext:felogin and ext:frontend. The table definitions need to be merged into one table which is
done by \TYPO3\CMS\Install\Service\SqlExpectedSchemaService::flattenSchemas().

The class also got a signal method emitTablesDefinitionIsBeingBuiltSignalDoctrine()
for the signal/slot mechanism.

The classes \TYPO3\CMS\Core\Cache\Schema\Typo3DatabaseBackendCacheSchema and
\TYPO3\CMS\Core\Cache\Schema\Typo3DatabaseBackendTagsSchema creating
the database schemas for the caching tables on-the-fly. While this is
easy for nearly all of the tables, it got tricky when it came to caching
tables for extbase because they were created by TYPO3 CMS with a random
number first. Luckily the \Doctrine\DBAL\Schema\Schema provide a method
to rename the schema which happen in TYPO3\CMS\Install\Service\CachingFrameworkDatabaseSchemaService::getCachingFrameworkRequiredDatabaseSchemaDoctrine().

To get the current database schema the extension XLCASSes
\TYPO3\CMS\Install\Service\SqlSchemaMigrationService and implement
the method \Konafets\DoctrineDbal\Install\Service\SqlSchemaMigrationService::getCurrentSchemaFromDatabase()
which ask Doctrine DBAL SchemaManager for the current database schema.

The schema manager determines the difference between the expected and the
current schema and returns the SQL query with the tables definitions
which is send by \DoctrineDbal\Install\Service\SqlSchemaMigrationService::performUpdateQueries()
to the database.